# backends/catalyst-interviewai/openapi-spec.yaml
swagger: '2.0'
info:
  title: InterviewAI Catalyst Backend API
  description: API for managing user-provided API keys, proxying AI requests, and RAG content processing for InterviewAI.
  version: '2.0.0'
# The host will be automatically generated when you deploy your API Gateway.
# Example: catalyst-gateway-xxxxxx-uc.a.run.app
# You can update this field later after deployment if needed.
host: your-gateway-id.uc.gateway.dev 
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json

# Disable all security - let Cloud Functions handle authentication
security: []
paths:
...
  /api/vector/search:
    options:
      summary: Handle CORS preflight requests for Vector Search
      operationId: corsVectorSearch
      security: []
      x-google-backend:
        address: "%s" # Placeholder for VectorSearch URL (15th)
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    post:
      summary: Performs semantic search using vector similarity
      description: Advanced semantic search with embeddings
      operationId: vectorSearch
      security: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - query
            properties:
              query:
                type: string
                description: Search query
              filters:
                type: object
                description: Metadata filters
              limit:
                type: integer
                default: 10
                description: Maximum results
      x-google-backend:
        address: "%s" # Placeholder for VectorSearch URL (16th)
        disable_auth: true
      responses:
        '200':
          description: Search results with relevance scores
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: '#/definitions/SearchResult'
              total:
                type: integer
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /api/vector/upsert:
    options:
      summary: Handle CORS preflight requests for Vector Upsert
      operationId: corsVectorUpsert
      security: []
      x-google-backend:
        address: "%s" # Placeholder for VectorSearch URL (17th)
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    post:
      summary: Upserts document embeddings to the vector index
      description: Stores or updates embeddings for vector search
      operationId: upsertVectorEmbeddings
      security: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - documents
            properties:
              documents:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    content:
                      type: string
                    embeddings:
                      type: object
                    metadata:
                      type: object
      x-google-backend:
        address: "%s" # Placeholder for VectorSearch URL (18th)
        disable_auth: true
      responses:
        '200':
          description: Embeddings upserted successfully
          schema:
            type: object
            properties:
              success:
                type: boolean
              upserted:
                type: integer
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /api/vector/similar:
    options:
      summary: Handle CORS preflight requests for Similar Content
      operationId: corsSimilarContent
      security: []
      x-google-backend:
        address: "%s" # Placeholder for VectorSearch URL (19th)
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    get:
      summary: Finds content similar to a given document
      description: Returns similar content based on vector similarity
      operationId: findSimilarContent
      security: []
      parameters:
        - name: documentId
          in: query
          type: string
          required: true
          description: ID of the reference document
        - name: limit
          in: query
          type: integer
          default: 5
          description: Maximum number of similar documents
      x-google-backend:
        address: "%s" # Placeholder for VectorSearch URL (20th)
        disable_auth: true
      responses:
        '200':
          description: Similar content found
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: '#/definitions/SearchResult'
              documentId:
                type: string
              total:
                type: integer
        '400':
          description: Bad request
        '401':
          description: Unauthorized

  # ===== PYTHON AGENT ENDPOINTS =====
...
definitions:
...
  SearchResult:
    type: object
    properties:
      id:
        type: string
      content:
        type: string
      source:
        type: string
      title:
        type: string
      score:
        type: number
      contentType:
        type: string
      metadata:
        type: object
...